+++ History.md
@@ -1,4 +1,7 @@
 # History
+## 2019-08-29 Version  1.0.8 @unclexiao
+- 删除基本框架[egg-tell-basic](https://www.npmjs.com/package/egg-tell-basic)
+- 保留框架的[初始配置](/boilerplate/config/config.default.js)
 
 ## 2019-08-12 Version  1.0.4 @unclexiao
 - 支持[多环境配置](https://eggjs.org/zh-cn/basics/config.html)：单元测试、开发环境、生产环境